Move the scheduler code to a separate file.
In this way it is easier to maintain.
